// src/model/address.ts
import { Address } from "@prisma/client";

export type ResponseAddress = {
  id: number;
  street?: string | null;
  city?: string | null;
  province?: string | null;
  country: string;
  postal_code: string;
};

export type GetAddressRequest = {
  id: number;
  contact_id: number;
};

export type CreateRequestAddress = {
  contact_id: number;
  street?: string;
  city?: string;
  province?: string;
  country: string;
  postal_code: string;
};

export type UpdateRequestAddress = {
  id: number;
  contact_id: number;
  street?: string;
  city?: string;
  province?: string;
  country: string;
  postal_code: string;
};

export function toAddressResponse(address: Address): ResponseAddress {
  return {
    id: address.id,
    street: address.street,
    city: address.city,
    province: address.province,
    country: address.country,
    postal_code: address.postal_code,
  };
}